[WFCORE-2946]: Unifiying the 3 FileStreamEntry definitions
