+++ controller-client/src/main/java/org/jboss/as/controller/client/helpers/standalone/impl/DeploymentPlanBuilderImpl.java
@@ -22,16 +22,12 @@
 
 package org.jboss.as.controller.client.helpers.standalone.impl;
 
-import java.io.BufferedInputStream;
-import java.io.DataOutput;
 import java.io.File;
-import java.io.FilterInputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.net.URISyntaxException;
 import java.net.URL;
 import java.net.URLConnection;
-import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.Collections;
@@ -48,7 +44,7 @@ import org.jboss.as.controller.client.helpers.standalone.InitialDeploymentPlanBu
 import org.jboss.as.controller.client.helpers.standalone.ReplaceDeploymentPlanBuilder;
 import org.jboss.as.controller.client.helpers.standalone.UndeployDeploymentPlanBuilder;
 import org.jboss.as.controller.client.helpers.standalone.DeploymentAction.Type;
-import org.jboss.as.controller.client.impl.InputStreamEntry;
+import org.jboss.as.controller.client.impl.InputStreamEntry.FileStreamEntry;
 import org.jboss.as.protocol.StreamUtils;
 
 /**
@@ -448,34 +444,6 @@ class DeploymentPlanBuilderImpl
     public DeploymentPlanBuilder removeContenFromDeployment(String deploymentName, List<String> paths) throws IOException {
          DeploymentActionImpl mod = DeploymentActionImpl.getRemoveContentAction(deploymentName, paths);
         return new DeploymentPlanBuilderImpl(this, mod);
-    }
-
-    // Wrap the FIS in a streamEntry so that the controller-client has access to the underlying File
-    private static class FileStreamEntry extends FilterInputStream implements InputStreamEntry {
-
-        private final Path file;
-
-        private FileStreamEntry(final File file) throws IOException {
-            this(file.toPath());
-        }
-
-        private FileStreamEntry(final Path file) throws IOException {
-            super(Files.newInputStream(file)); // This stream will get closed regardless of autoClose
-            this.file = file;
-        }
-
-        @Override
-        public int initialize() throws IOException {
-            return (int) Files.size(file);
-        }
-
-        @Override
-        public void copyStream(final DataOutput output) throws IOException {
-            try (InputStream in = new BufferedInputStream(Files.newInputStream(file))) {
-                StreamUtils.copyStream(in, output);
-            }
-        }
-
     }
 
 }
